Time: 36 ms (17.84%), Space: 12.1 MB (78.25%) - LeetHub

+++ 0150-evaluate-reverse-polish-notation/0150-evaluate-reverse-polish-notation.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int evalRPN(vector<string>& tokens) {
         stack<int> st;
-        long long a, b;
+        long a, b;
         for(auto t: tokens) {
             if(t == "+" or t == "-" or t == "*" or t == "/") {
                 b = st.top(); st.pop();
